header_size needs to be able to handle MAX_HEADER_LENGTH=65535 bytes, so make it an int instead of a short
git-svn-id: svn://svn.code.sf.net/p/hfst/code/trunk/hfst3@2531 941e2c2b-deac-454f-805a-451daa25f33c

+++ libhfst/src/HfstInputStream.cc
@@ -775,7 +775,7 @@ namespace hfst
 
   int HfstInputStream::get_header_size(int &bytes_read)
   {
-    short header_size=0;
+    int header_size=0;
     stream_get(header_size);
     char c = stream_get();
     if (c != 0) {
